Re-Implement CI for the repo, plus some housekeeping
This uses deployments to keep runners only using specific credentials
for staging and production. Proper handling of staging so people can
test deploy shit will come later.

This also uses DHCP on wiggle, because we're not using a public IP for
it any longer.

Rotates the key colmena_ci uses too, plus adds a staging version of the
module for uh, staging.

Finally, removes frankfurt 2 and 3, plus fixes the pkgs-unstable
reference on the arm nodes to use arm.

Co-authored-by: alexkar598
<25136265+alexkar598@users.noreply.github.com>